# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample Server pets",
        "title": "Swagger Example API",
        "contact": {},
        "version": "0.0.1"
    },
    "basePath": "/",
    "paths": {
...
        "/project/createProject": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Project"
                ],
                "summary": "创建Project",
                "parameters": [
                    {
                        "description": "创建Project",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/autocode.Project"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"获取成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/project/deleteProject": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Project"
                ],
                "summary": "删除Project",
                "parameters": [
                    {
                        "description": "删除Project",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/autocode.Project"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"删除成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/project/deleteProjectByIds": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Project"
                ],
                "summary": "批量删除Project",
                "parameters": [
                    {
                        "description": "批量删除Project",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.IdsReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"批量删除成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/project/findProject": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Project"
                ],
                "summary": "用id查询Project",
                "parameters": [
                    {
                        "type": "string",
                        "name": "area",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "name": "cAmount",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "name": "cRadio",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "categories",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "chargeStandard",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "contractEndDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "contractNum",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "contractStartDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "创建时间",
                        "name": "createdAt",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "createdDate",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "name": "dAmount",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "name": "dRadio",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "主键ID",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "manager",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "name": "paidAmount",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "name": "projectAmount",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "projectCode",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "remark",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "name": "sAmount",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "name": "sRadio",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "trainEndDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "trainMode",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "trainNumOfPerson",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "trainStartDate",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "trainTime",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "name": "unpaidAmount",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "更新时间",
                        "name": "updatedAt",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "name": "wAmount",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "name": "wRadio",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"查询成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/project/getProjectList": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Project"
                ],
                "summary": "分页获取Project列表",
                "parameters": [
                    {
                        "type": "string",
                        "name": "area",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "name": "cAmount",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "name": "cRadio",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "categories",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "chargeStandard",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "contractEndDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "contractNum",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "contractStartDate",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "name": "contractStartDateRange",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "创建时间",
                        "name": "createdAt",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "createdDate",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "这里是为了日期范围筛选添加的几个日期数组字段",
                        "name": "createdDateRange",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "name": "dAmount",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "name": "dRadio",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "主键ID",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "manager",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页大小",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "name": "paidAmount",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "name": "projectAmount",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "projectCode",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "remark",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "name": "sAmount",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "name": "sRadio",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "trainEndDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "trainMode",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "trainNumOfPerson",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "trainStartDate",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "name": "trainStartDateRange",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "trainTime",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "name": "unpaidAmount",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "更新时间",
                        "name": "updatedAt",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "name": "wAmount",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "name": "wRadio",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"获取成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/project/updateProject": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Project"
                ],
                "summary": "更新Project",
                "parameters": [
                    {
                        "description": "更新Project",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/autocode.Project"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"更新成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "autocode.AmountObj": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "radio": {
                    "type": "number"
                }
            }
        },
...
        "autocode.IncomeAndOutcomeObj": {
            "type": "object",
            "properties": {
                "pg": {
                    "type": "number"
                },
                "ph": {
                    "type": "number"
                },
                "pi": {
                    "type": "number"
                },
                "pj": {
                    "type": "number"
                },
                "pk": {
                    "type": "number"
                },
                "pl": {
                    "type": "number"
                },
                "pm": {
                    "type": "number"
                },
                "pn": {
                    "type": "number"
                },
                "po": {
                    "type": "number"
                },
                "pp": {
                    "type": "number"
                },
                "pq": {
                    "type": "number"
                },
                "pr": {
                    "type": "number"
                },
                "ps": {
                    "type": "number"
                },
                "pt": {
                    "type": "number"
                }
            }
        },
...
        "autocode.Project": {
            "type": "object",
            "required": [
                "client",
                "createdDate",
                "incomeAndOutcome",
                "landingAgency",
                "partner"
            ],
            "properties": {
                "area": {
                    "type": "string"
                },
                "cAmount": {
                    "type": "number"
                },
                "cRadio": {
                    "type": "number"
                },
                "categories": {
                    "type": "string"
                },
                "chargeStandard": {
                    "type": "string"
                },
                "client": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/autocode.AmountObj"
                    }
                },
                "contractEndDate": {
                    "type": "string"
                },
                "contractNum": {
                    "type": "string"
                },
                "contractStartDate": {
                    "type": "string"
                },
                "createdAt": {
                    "description": "创建时间",
                    "type": "string"
                },
                "createdDate": {
                    "type": "string"
                },
                "dAmount": {
                    "type": "number"
                },
                "dRadio": {
                    "type": "number"
                },
                "id": {
                    "description": "主键ID",
                    "type": "integer"
                },
                "incomeAndOutcome": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/autocode.IncomeAndOutcomeObj"
                    }
                },
                "landingAgency": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/autocode.AmountObj"
                    }
                },
                "manager": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "paidAmount": {
                    "type": "number"
                },
                "partner": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/autocode.AmountObj"
                    }
                },
                "projectAmount": {
                    "type": "number"
                },
                "projectCode": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "sAmount": {
                    "type": "number"
                },
                "sRadio": {
                    "type": "number"
                },
                "trainEndDate": {
                    "type": "string"
                },
                "trainMode": {
                    "type": "string"
                },
                "trainNumOfPerson": {
                    "type": "integer"
                },
                "trainStartDate": {
                    "type": "string"
                },
                "trainTime": {
                    "type": "integer"
                },
                "unpaidAmount": {
                    "type": "number"
                },
                "updatedAt": {
                    "description": "更新时间",
                    "type": "string"
                },
                "wAmount": {
                    "type": "number"
                },
                "wRadio": {
                    "type": "number"
                }
            }
        },
...
        "request.IdsReq": {
            "type": "object",
            "properties": {
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "x-token",
            "in": "header"
        }
    }
}